streamlit-rag-chatbot/requirements.txt
streamlit
llama-index
llama-cloud-services
python-dotenv
httpx 